Don't remove verticies in `Search.update/2`.
Previously, the update function removed any verticies not reachable from
the new root vertex. This was very slow due to the naive pathfinding,
and realistically it's probably not necessary to remove them anyway if
the root vertex ID is updated.

In the future it may be worth trying it again with a more performance
implementation, but for now we'll trade graph size/memory for speed.